[REFACTOR]: Update routes dir
