+++ src/infra/lib/hero.ts
