+++ src/interfaces/interfaceProps.ts
@@ -1,5 +1,5 @@
-import {Contact} from '@lib/contacts';
-import {Link} from '@lib/links';
+import {Contact} from '../infra/lib/contacts';
+import {Link} from '../infra/lib/links';
 import {StaticImageData} from 'next/image';
 import {IconType} from 'react-icons';
 import {ReactNode} from 'react';
