+++ src/patterns/Carousel/Button.tsx
@@ -1,5 +1,5 @@
 import {ButtonProps} from '@interfaces/interfaceProps';
-import Link from '@components/Link';
+import Link from '../../components/Link';
 
 export default function Button({ linkProject }: ButtonProps) {
   return (
