+++ src/patterns/Carousel/Card.tsx
