+++ src/patterns/Carousel/index.tsx
@@ -1,4 +1,4 @@
-import {projects} from '@lib/projects';
+import {projects} from '../../infra/lib/projects';
 import {Carousel} from 'react-responsive-carousel';
 import 'react-responsive-carousel/lib/styles/carousel.min.css';
 import Card from './Card';
