+++ src/patterns/Content.tsx
