+++ src/patterns/ContentAboutMe.tsx
