+++ src/patterns/Footer/Copyright.tsx
@@ -1,4 +1,4 @@
-import Link from '@components/Link';
+import Link from '../../components/Link';
 
 export default function Copyright() {
   return (
