+++ src/patterns/Hero.tsx
