+++ src/patterns/NavBar/Links.tsx
@@ -1,5 +1,5 @@
 import {LinksProps} from '@interfaces/interfaceProps';
-import Link from '@components/Link';
+import Link from '../../components/Link';
 
 export default function Links({
   links,
