+++ src/patterns/NavBar/Menu.tsx
@@ -1,4 +1,4 @@
-import {AppContext} from '../../Layout';
+import {AppContext} from '@components/Layout';
 import {MenuProps} from '@interfaces/interfaceProps';
 import {useContext} from 'react';
 import {FaBars, FaTimes} from 'react-icons/fa';
